# contracts/openapi.yaml
...
openapi: 3.1.0
...
info:
  title: VN Terminal Pro API
  version: 0.1.0
  description: Local, single-user API for VN Terminal Pro (Section 11 of vn_terminal_multi_ai_development_plan.md
    v1.1). Binds to 127.0.0.1 only (Section 3.1, Section 20); never exposed publicly. This file is generated
    by contracts/build_openapi.py from the Pydantic models in contracts/schemas/models/. Do not hand-edit.
servers:
- url: http://127.0.0.1:{port}
  description: Local-only backend (Section 3.1).
  variables:
    port:
      default: '8000'
...
paths:
...
  /api/data/refresh:
    post:
      operationId: startDataRefresh
      summary: Start atomic data refresh
      tags:
      - data
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataRefreshRequest'
      responses:
        '202':
          description: Refresh accepted and queued/running.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataRefreshResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/data/status:
    get:
      operationId: getDataStatus
      summary: Refresh progress and latest good snapshot
      tags:
      - data
      responses:
        '200':
          description: Refresh status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataStatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/symbols:
    get:
      operationId: listSymbols
      summary: Search and filter the security universe
      tags:
      - data
      parameters:
      - $ref: '#/components/parameters/SymbolSearchQuery'
      - $ref: '#/components/parameters/ExchangeFilterQuery'
      - $ref: '#/components/parameters/SecurityTypeFilterQuery'
      - $ref: '#/components/parameters/TradingStatusFilterQuery'
      - $ref: '#/components/parameters/LimitQuery'
      - $ref: '#/components/parameters/OffsetQuery'
      responses:
        '200':
          description: Matching securities.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SymbolsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/bars/{symbol}:
    get:
      operationId: getBars
      summary: Return validated OHLCV
      tags:
      - data
      parameters:
      - $ref: '#/components/parameters/SymbolPath'
      - $ref: '#/components/parameters/TimeframeQuery'
      - $ref: '#/components/parameters/StartDateQuery'
      - $ref: '#/components/parameters/EndDateQuery'
      responses:
        '200':
          description: Validated OHLCV bars.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BarsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/fundamentals/{symbol}:
    get:
      operationId: getFundamentals
      summary: Financial statements and derived metrics
      tags:
      - fundamentals
      parameters:
      - $ref: '#/components/parameters/SymbolPath'
      responses:
        '200':
          description: Financial observations and derived metrics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FundamentalsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    BarsResponse:
      additionalProperties: false
      properties:
        symbol:
          title: Symbol
          type: string
        timeframe:
          $ref: '#/components/schemas/Timeframe'
        bars:
          items:
            $ref: '#/components/schemas/PriceBar'
          title: Bars
          type: array
        provenance:
          $ref: '#/components/schemas/Provenance'
      required:
      - symbol
      - timeframe
      - bars
      - provenance
      title: BarsResponse
      type: object
    DataRefreshRequest:
      additionalProperties: false
      properties:
        force:
          default: false
          description: 'OPEN ITEM: not named by the plan. Added so a caller can request a refresh even
            if one recently succeeded. See contracts/OPEN_ITEMS.md.'
          title: Force
          type: boolean
      title: DataRefreshRequest
      type: object
    DataRefreshResponse:
      additionalProperties: false
      properties:
        run_id:
          title: Run Id
          type: string
        status:
          description: 'OPEN ITEM: see contracts/OPEN_ITEMS.md, ''Job/health status vocabulary.'''
          examples:
          - queued
          - running
          title: Status
          type: string
        started_at:
          format: date-time
          title: Started At
          type: string
        provenance:
          $ref: '#/components/schemas/Provenance'
      required:
      - run_id
      - status
      - started_at
      - provenance
      title: DataRefreshResponse
      type: object
    DataStatusResponse:
      additionalProperties: false
      properties:
        run_id:
          anyOf:
          - type: string
          - type: 'null'
          default: null
          title: Run Id
        status:
          description: 'OPEN ITEM: see contracts/OPEN_ITEMS.md, ''Job/health status vocabulary.'''
          examples:
          - queued
          - running
          - completed
          - failed
          title: Status
          type: string
        progress_pct:
          anyOf:
          - maximum: 100
            minimum: 0
            type: number
          - type: 'null'
          default: null
          title: Progress Pct
        latest_good_snapshot:
          anyOf:
          - $ref: '#/components/schemas/SnapshotInfo'
          - type: 'null'
          default: null
          description: 'Section 3.4: ''The previous validated data snapshot remains available if a refresh
            fails.'''
        issues:
          items:
            type: string
          title: Issues
          type: array
        provenance:
          $ref: '#/components/schemas/Provenance'
      required:
      - status
      - provenance
      title: DataStatusResponse
      type: object
...
    DerivedMetric:
      additionalProperties: false
      description: 'Section 17: ''Calculate growth and margins deterministically.''


        OPEN ITEM: the plan lists panel content (Section 17, "Initial panels")

        but not a literal derived-metric record shape. This mirrors

        FinancialObservation''s shape closely so both can be rendered the same

        way in the fundamentals panel (Section 12.3).'
      properties:
        symbol:
          title: Symbol
          type: string
        period_end:
          format: date
          title: Period End
          type: string
        metric_code:
          examples:
          - revenue_yoy_growth
          - gross_margin
          - roe
          - roa
          - pe
          - pb
          title: Metric Code
          type: string
        metric_label_vi:
          title: Metric Label Vi
          type: string
        value:
          title: Value
          type: number
        unit:
          title: Unit
          type: string
      required:
      - symbol
      - period_end
      - metric_code
      - metric_label_vi
      - value
      - unit
      title: DerivedMetric
      type: object
...
    ErrorResponse:
      additionalProperties: false
      description: Explicit error response (Section 11) for non-validation errors.
      properties:
        error_code:
          description: 'OPEN ITEM: the plan requires ''explicit error responses'' (Section 11) but does
            not name an error vocabulary. See contracts/OPEN_ITEMS.md.'
          title: Error Code
          type: string
        message:
          title: Message
          type: string
        details:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          default: null
          title: Details
      required:
      - error_code
      - message
      title: ErrorResponse
      type: object
...
    FinancialObservation:
      additionalProperties: false
      properties:
        symbol:
          title: Symbol
          type: string
        period_end:
          format: date
          title: Period End
          type: string
        period_type:
          $ref: '#/components/schemas/PeriodType'
        statement_type:
          description: 'OPEN ITEM: Section 10.3 names the field but does not enumerate statement types.
            Examples given here follow standard financial-statement categories implied by Section 17 (''Financial
            statements'') but are not a plan-given closed list. See contracts/OPEN_ITEMS.md.'
          examples:
          - income_statement
          - balance_sheet
          - cash_flow_statement
          title: Statement Type
          type: string
        metric_code:
          title: Metric Code
          type: string
        metric_label_vi:
          title: Metric Label Vi
          type: string
        value:
          title: Value
          type: number
        currency:
          examples:
          - VND
          title: Currency
          type: string
        unit:
          description: 'OPEN ITEM: Section 17 requires recording ''currency and units'' but the plan does
            not give a unit vocabulary (for example whether values are in VND, thousand VND, or million
            VND). See contracts/OPEN_ITEMS.md.'
          title: Unit
          type: string
        consolidated:
          description: 'Section 17: ''Distinguish consolidated and separate statements.'''
          title: Consolidated
          type: boolean
        restatement_version:
          description: 'OPEN ITEM: Section 10.3 names ''restatement_version'' and Section 17 requires
            ''Preserve restatements'' but the plan does not say whether this is an incrementing integer
            or a string tag. Modeled as an integer counter starting at 1 in this session. See contracts/OPEN_ITEMS.md.'
          minimum: 1
          title: Restatement Version
          type: integer
        publication_date:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          default: null
          description: 'Section 17 degraded mode: nullable. When null, this observation must be excluded
            from historical backtests (Section 16: ''No use of later financial-publication information...
            affected fundamental inputs are excluded from backtests rather than assumed'') and must carry
            an explicit warning at the point of use.'
          title: Publication Date
        source_url:
          title: Source Url
          type: string
        ingested_at:
          format: date-time
          title: Ingested At
          type: string
      required:
      - symbol
      - period_end
      - period_type
      - statement_type
      - metric_code
      - metric_label_vi
      - value
      - currency
      - unit
      - consolidated
      - restatement_version
      - source_url
      - ingested_at
      title: FinancialObservation
      type: object
    FundamentalsResponse:
      additionalProperties: false
      properties:
        symbol:
          title: Symbol
          type: string
        observations:
          items:
            $ref: '#/components/schemas/FinancialObservation'
          title: Observations
          type: array
        derived_metrics:
          items:
            $ref: '#/components/schemas/DerivedMetric'
          title: Derived Metrics
          type: array
        provenance:
          $ref: '#/components/schemas/Provenance'
      required:
      - symbol
      - observations
      - provenance
      title: FundamentalsResponse
      type: object
    HTTPValidationError:
      additionalProperties: false
      description: 'FastAPI''s standard 422 request-validation error envelope.


        Traceable to Section 7 (''Backend: Python FastAPI''); this is FastAPI''s

        own documented default shape, not an invented one.'
      properties:
        detail:
          items:
            $ref: '#/components/schemas/HTTPValidationErrorItem'
          title: Detail
          type: array
      required:
      - detail
      title: HTTPValidationError
      type: object
    HTTPValidationErrorItem:
      additionalProperties: false
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Loc
          type: array
        msg:
          title: Msg
          type: string
        type:
          title: Type
          type: string
      required:
      - loc
      - msg
      - type
      title: HTTPValidationErrorItem
      type: object
...
    PeriodType:
      description: 'Section 10.3: "period_type, quarter or year.'
      enum:
      - quarter
      - year
      title: PeriodType
      type: string
    PriceBar:
      additionalProperties: false
      properties:
        symbol:
          examples:
          - FPT
          title: Symbol
          type: string
        exchange:
          description: 'Section 1/10.1: evidenced values are HOSE, HNX, UPCoM. Left as an open string
            rather than a closed enum; see contracts/OPEN_ITEMS.md.'
          examples:
          - HOSE
          - HNX
          - UPCoM
          title: Exchange
          type: string
        security_type:
          description: 'OPEN ITEM: candidate values inferred from WP1''s four FData category directories
            (stock, index, der, cw) mapping to equity/index/derivative/warrant, plus ''bond'' per the
            Section 3.2 assumption text, which has no confirmed FData source directory. Section 19.2 also
            notes ''Mixed security types in the stock directory,'' so directory alone does not determine
            this value. Left as an open string; see contracts/OPEN_ITEMS.md.'
          examples:
          - equity
          title: Security Type
          type: string
        timeframe:
          $ref: '#/components/schemas/Timeframe'
          examples:
          - 1D
        trading_date:
          format: date
          title: Trading Date
          type: string
        timezone:
          default: Asia/Ho_Chi_Minh
          description: Section 10.1 example value; VN Terminal Pro is single-market.
          title: Timezone
          type: string
        open:
          description: Thousands of VND, back-adjusted (Section 19.1).
          title: Open
          type: number
        high:
          description: Thousands of VND, back-adjusted (Section 19.1).
          title: High
          type: number
        low:
          description: Thousands of VND, back-adjusted (Section 19.1).
          title: Low
          type: number
        close:
          description: Thousands of VND, back-adjusted (Section 19.1).
          title: Close
          type: number
        volume:
          description: Shares (Section 19.1).
          minimum: 0
          title: Volume
          type: integer
        adjustment_status:
          description: 'OPEN ITEM: only ''adjusted_unknown_method'' is evidenced (Section 10.1 example).
            The plan does not enumerate the full set of possible values. See contracts/OPEN_ITEMS.md.'
          examples:
          - adjusted_unknown_method
          title: Adjustment Status
          type: string
        source:
          examples:
          - Fialda FData
          title: Source
          type: string
        source_file:
          examples:
          - C:\FDATA\AmiBroker\EOD\stock\FPT.dat
          title: Source File
          type: string
        ingested_at:
          format: date-time
          title: Ingested At
          type: string
        quality_status:
          description: 'OPEN ITEM: only ''valid'' is evidenced (Section 10.1 example). See contracts/OPEN_ITEMS.md.'
          examples:
          - valid
          title: Quality Status
          type: string
      required:
      - symbol
      - exchange
      - security_type
      - timeframe
      - trading_date
      - open
      - high
      - low
      - close
      - volume
      - adjustment_status
      - source
      - source_file
      - ingested_at
      - quality_status
      title: PriceBar
      type: object
    Provenance:
      additionalProperties: false
      description: Attached to every API response envelope (Section 11, Section 3.4).
      properties:
        as_of_date:
          description: 'Section 11: ''as_of_date'' required on every endpoint.'
          format: date
          title: As Of Date
          type: string
        data_version:
          description: 'Section 11: ''data_version'' required on every endpoint. Example form from Section
            10.5: ''fdata-2026-07-30''.'
          title: Data Version
          type: string
        source:
          description: 'Section 11: ''source'' required on every endpoint.'
          title: Source
          type: string
        freshness_status:
          description: 'Section 3.4: ''freshness status''. OPEN ITEM: the plan does not enumerate allowed
            values. See contracts/OPEN_ITEMS.md.'
          title: Freshness Status
          type: string
        warnings:
          description: 'Section 11: ''Quality and freshness warnings where applicable.'''
          items:
            type: string
          title: Warnings
          type: array
      required:
      - as_of_date
      - data_version
      - source
      - freshness_status
      title: Provenance
      type: object
...
    SecurityMaster:
      additionalProperties: false
      properties:
        symbol:
          title: Symbol
          type: string
        isin:
          anyOf:
          - type: string
          - type: 'null'
          default: null
          description: 'Section 10.2: ''isin, when available.'''
          title: Isin
        exchange:
          description: See OPEN_ITEMS.md; same open-string treatment as PriceBar.exchange.
          examples:
          - HOSE
          - HNX
          - UPCoM
          title: Exchange
          type: string
        security_type:
          description: See OPEN_ITEMS.md; same open-string treatment as PriceBar.security_type.
          examples:
          - equity
          title: Security Type
          type: string
        company_name:
          title: Company Name
          type: string
        sector:
          title: Sector
          type: string
        industry:
          title: Industry
          type: string
        listing_date:
          format: date
          title: Listing Date
          type: string
        delisting_date:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          default: null
          description: Section 10.2 lists this field; only populated for delisted securities.
          title: Delisting Date
        trading_status:
          description: 'OPEN ITEM: Section 3.2 mentions distinguishing ''suspended securities, and delisted
            securities''; WP2 (Section 22) speaks of ''Active/inactive status.'' Whether trading_status
            is a three-plus-value set (active/suspended/delisted/...) or a two-value active/inactive flag
            with suspension detail held elsewhere is not settled by the plan. See contracts/OPEN_ITEMS.md.'
          examples:
          - active
          - suspended
          - delisted
          title: Trading Status
          type: string
        currency:
          examples:
          - VND
          title: Currency
          type: string
        price_unit:
          description: 'Section 3.2: ''OHLC prices appear to be historically adjusted and expressed in
            thousands of VND.'''
          examples:
          - thousand_vnd
          title: Price Unit
          type: string
        lot_size:
          minimum: 1
          title: Lot Size
          type: integer
        source:
          title: Source
          type: string
        valid_from:
          format: date
          title: Valid From
          type: string
        valid_to:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          default: null
          title: Valid To
        index_code:
          anyOf:
          - type: string
          - type: 'null'
          default: null
          description: Only populated for security_type=index entries sourced from EOD/index, whose 207
            files are named with numeric codes (for example 0001.dat, 0500.dat) rather than symbols (Section
            10.2). Not a literal field name from the plan; see contracts/OPEN_ITEMS.md.
          title: Index Code
      required:
      - symbol
      - exchange
      - security_type
      - company_name
      - sector
      - industry
      - listing_date
      - trading_status
      - currency
      - price_unit
      - lot_size
      - source
      - valid_from
      title: SecurityMaster
      type: object
...
    SnapshotInfo:
      additionalProperties: false
      properties:
        data_version:
          title: Data Version
          type: string
        as_of_date:
          format: date
          title: As Of Date
          type: string
        promoted_at:
          format: date-time
          title: Promoted At
          type: string
      required:
      - data_version
      - as_of_date
      - promoted_at
      title: SnapshotInfo
      type: object
...
    SymbolsResponse:
      additionalProperties: false
      properties:
        items:
          items:
            $ref: '#/components/schemas/SecurityMaster'
          title: Items
          type: array
        total:
          minimum: 0
          title: Total
          type: integer
        provenance:
          $ref: '#/components/schemas/Provenance'
      required:
      - items
      - total
      - provenance
      title: SymbolsResponse
      type: object
    Timeframe:
      description: 'Daily, weekly, monthly per Section 12.1 and Section 12.2.


        OPEN ITEM: the plan never gives literal timeframe codes. "1D" appears

        once, as an example value in the Section 10.1 price-bar JSON sample.

        "1W" and "1M" are inferred by this session to cover the weekly and

        monthly views required by Section 12.1/12.2. See contracts/OPEN_ITEMS.md.'
      enum:
      - 1D
      - 1W
      - 1M
      title: Timeframe
      type: string
...
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: The requested resource (symbol, run_id, watchlist id) does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Request failed schema validation (FastAPI default 422 shape, Section 7).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HTTPValidationError'
    ServiceUnavailable:
      description: 'Required data or an AI provider is unavailable, or the latest snapshot failed validation
        (Section 3.4: ''No trading recommendation is generated when required data is missing or fails
        validation.'').'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  parameters:
    SymbolPath:
      name: symbol
      in: path
      required: true
      schema:
        type: string
      example: FPT
...
    TimeframeQuery:
      name: timeframe
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/Timeframe'
      description: 'Section 12.1/12.2: daily, weekly, or monthly. Defaults to 1D.'
    StartDateQuery:
      name: start_date
      in: query
      required: false
      schema:
        type: string
        format: date
    EndDateQuery:
      name: end_date
      in: query
      required: false
      schema:
        type: string
        format: date
    SymbolSearchQuery:
      name: q
      in: query
      required: false
      schema:
        type: string
      description: 'Section 11: ''Search and filter the security universe.'''
    ExchangeFilterQuery:
      name: exchange
      in: query
      required: false
      schema:
        type: string
    SecurityTypeFilterQuery:
      name: security_type
      in: query
      required: false
      schema:
        type: string
    TradingStatusFilterQuery:
      name: trading_status
      in: query
      required: false
      schema:
        type: string
    LimitQuery:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
      description: 'OPEN ITEM: pagination is not specified by Section 11; added given the full-market
        universe is 1,500-1,800 securities (Section 3.3). See contracts/OPEN_ITEMS.md.'
    OffsetQuery:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
...